code/requirements.txt
fastapi
pydantic
requests
python-dotenv
python-multipart
uvicorn